Reverted menu changes
I reverted the changes I made to the menu regarding the about, restart, and quit functions. It's probably best to try to diagnose the problem with SAVE.SC instead.
Also, there is now a batch file that automatically allows the game to start with the -d parameter, for debugging purposes.