Updated source scripts
I have managed to address a graphical glitch regarding the Menu script after restoring a game. Basically, I put a space before the name of each menu item. That was done in Codename Iceman and KQ1SCI.
Also, I've changed the names of some of the globals and variables to more closely match their original names.