Some scripts updated
The PrintD procedure in INTERFACE.SC has been renamed to PrintIcon to differentiate it from the PrintD procedure in PRINTD.SC

Also, SAVE.SC's GetDirectory procedure has been reimplemented in script code, replacing the assembly code that the decompiler fell back to. As a result, TEXT.990 is no longer needed.