Added in palette.999 and updated Save script
It turns out, the EGA version of Seasoned professional actually has the VGA palette file! In fact,
the EGA and VGA versions are EXACTLY the same in terms of sctipts.  The only differences are the interpreter, the graphic driver, and (of course) the pics and views.

Also, I have fixed a bug in SAVE.SC where pressing Escape in the save/restore dialog would not exit it, but instead cause it to give junk text or the top save-game description in the edit field. Turns out a return variable was changed from 0 to -1 by SCI1.1.