Moved system script strings into Text resources
In the system scripts (save MENU.SC, which is intended to be customized), any text strings in quotation marks have been moved to Text resources.

Perhaps SCICompanion should implement a feature that automatically generates Text resources from strings found in the scripts. This is what Sierra's original compiler did, and we should disassemble it to make Companion's compiler more accurate.