Lots of changes
Added GameInvItem subclass. This allows for inventory item descriptions to be stored in Text.003.

Filled in the blanks in DEBUG.SC.

Filled in the blanks in GAME.SC's save: and restore: methods.

Added defines for howFast values.

The unused GAME_INV define is now INVDESC. This is intended only for the new Text resource for item descriptions.

Many procedures have had their names changed.
	AnimateCast -> RedrawCast
	PrintDontHaveIt -> DontHave
	PrintAlreadyDoneThat -> AlreadyDone
	PrintCantDoThat -> CantDo
	PrintCantSeeThat -> CantSee
	AddManyObstacles -> InitAddToPics
	AddManyFeatures -> InitFeatures

The game-specific globals have been reshuffled a bit for better organization, and many have had their names changed..
	musicChannels -> numVoices
	colorCount -> numColors
	howFast -> machineSpeed
	detailLevel -> howFast
	curTextColor -> myTextColor
	curBackColor -> myBackColor

The GameVerbMessager class is now the verbWords instance.

A few classes are now preloaded at initialization to ensure that they stay in memory.