Fixed major verbMessager bug in USER>SC
For some reason, I'd forgotten to include a line of code that allows User's said: method to work with the verbMessager! So if you were wondering why doVerb wasn't working, this is the reason.

Also, I've greatly simplified the game's handleEvent:, specifically the "look at" inventory items" code, and added ddistinct Walk and StopWalk cyclers for the game's Ego.