Added workaround for VGA
When bringing up the save dialog, exiting it causes it to remain on the screen, creating a risk of crashing with "PackHandle failure" errors. This only occurs in the VGA interpreter; the EGA interpreter has no such problem.

I did a workaround of this issue, but it creates a less-severe issue of views popping into the dialog box after a Print message.

Strangely, the inventory dialog (which is implemented in a similar way) does not have this problem at all.