Now use original defines for ENTER, TAB, and BACKSPACE
Thanks to key ticks no longer being case-sensitive, this created a positive side-effect of them being usable in defines! As a result, the original Sierra defines for ENTER, TAB, and BACKSPACE in SYSTEM.SH are now used.

Hopefully, SCICompanion will add support for macro defines, so that all of Sierra's original defines can be used.